move minion metrics definition to pinot-common package
