+++ pinot-common/src/main/java/org/apache/pinot/common/metrics/MinionMeter.java
@@ -16,26 +16,40 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.minion.metrics;
+package org.apache.pinot.common.metrics;
 
 import org.apache.pinot.common.Utils;
 import org.apache.pinot.common.metrics.AbstractMetrics;
 
 
-public enum MinionTimer implements AbstractMetrics.Timer {
-  ;
+public enum MinionMeter implements AbstractMetrics.Meter {
+  HEALTH_CHECK_GOOD_CALLS("healthChecks", true),
+  HEALTH_CHECK_BAD_CALLS("healthChecks", true),
 
-  private final String _timerName;
+  NUMBER_TASKS_EXECUTED("tasks", false),
+  NUMBER_TASKS_COMPLETED("tasks", false),
+  NUMBER_TASKS_CANCELLED("tasks", false),
+  NUMBER_TASKS_FAILED("tasks", false),
+  NUMBER_TASKS_FATAL_FAILED("tasks", false);
+
+  private final String _meterName;
+  private final String _unit;
   private final boolean _global;
 
-  MinionTimer(boolean global) {
-    _timerName = Utils.toCamelCase(name().toLowerCase());
+  MinionMeter(String unit, boolean global) {
+    _meterName = Utils.toCamelCase(name().toLowerCase());
+    _unit = unit;
     _global = global;
   }
 
   @Override
-  public String getTimerName() {
-    return _timerName;
+  public String getMeterName() {
+    return _meterName;
+  }
+
+  @Override
+  public String getUnit() {
+    return _unit;
   }
 
   @Override
