+++ pinot-common/src/main/java/org/apache/pinot/common/metrics/MinionMetrics.java
@@ -16,38 +16,34 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.minion.metrics;
+package org.apache.pinot.common.metrics;
 
-import org.apache.pinot.common.Utils;
-import org.apache.pinot.common.metrics.AbstractMetrics;
+import org.apache.pinot.spi.metrics.PinotMetricsRegistry;
+import org.apache.pinot.spi.utils.CommonConstants;
 
 
-public enum MinionGauge implements AbstractMetrics.Gauge {
-  NUMBER_OF_TASKS("tasks", true),
-  ;
+public class MinionMetrics extends AbstractMetrics<MinionQueryPhase, MinionMeter, MinionGauge, MinionTimer> {
 
-  private final String _gaugeName;
-  private final String _unit;
-  private final boolean _global;
+  public MinionMetrics(PinotMetricsRegistry metricsRegistry) {
+    this(CommonConstants.Minion.CONFIG_OF_METRICS_PREFIX, metricsRegistry);
+  }
 
-  MinionGauge(String unit, boolean global) {
-    _gaugeName = Utils.toCamelCase(name().toLowerCase());
-    _unit = unit;
-    _global = global;
+  public MinionMetrics(String prefix, PinotMetricsRegistry metricsRegistry) {
+    super(prefix, metricsRegistry, MinionMetrics.class);
   }
 
   @Override
-  public String getGaugeName() {
-    return _gaugeName;
+  protected MinionQueryPhase[] getQueryPhases() {
+    return MinionQueryPhase.values();
   }
 
   @Override
-  public String getUnit() {
-    return _unit;
+  protected MinionMeter[] getMeters() {
+    return MinionMeter.values();
   }
 
   @Override
-  public boolean isGlobal() {
-    return _global;
+  protected MinionGauge[] getGauges() {
+    return MinionGauge.values();
   }
 }
