+++ pinot-common/src/main/java/org/apache/pinot/common/metrics/MinionQueryPhase.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.minion.metrics;
+package org.apache.pinot.common.metrics;
 
 import org.apache.pinot.common.Utils;
 import org.apache.pinot.common.metrics.AbstractMetrics;
